dssm initialize implementation

+++ d-ssm-app/src/main/java/com/dssm/mapper/manage/PermissionMapper.java
@@ -7,6 +7,6 @@ import com.dssm.mapper.BaseMapper;
 
 public interface PermissionMapper extends BaseMapper<Permission> {
     
-    public List<Permission> selectByRole(int roleId);
+    public List<Permission> selectByRole(Long roleId);
     
 }
